+++ hw12_13_14_15_calendar/internal/grpc/server/server.go
@@ -1,3 +1,4 @@
+//TODO: Rename to server_grpc
 package server
 
 import (
