+++ hw12_13_14_15_calendar/internal/logger/logger.go
@@ -15,14 +15,14 @@ type Instance struct {
 
 func (i *Instance) Init(path string) error {
 	// If the file doesn't exist, create it or append to the file
-	file, err := os.OpenFile(path, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0666)
-	if err != nil {
-		log.Fatal(err)
+	//file, err := os.OpenFile(path, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0666)
+	//if err != nil {
+	//	log.Fatal(err)
+	//
+	//	return err
+	//}
 
-		return err
-	}
-
-	log.SetOutput(file)
+	log.SetOutput(os.Stdout)
 	fmt.Println("logger initialized, log file: ", path)
 
 	return nil
